refactor: rename model_adapters module to adapters
