+++ src/adapters/AGENTS.md
