+++ src/adapters/CLAUDE.md
