+++ src/adapters/mod.rs
